Initial RaidFinder merge

+++ Source/Core/Gen5/Nazos.hpp
@@ -24,7 +24,7 @@
 #include <array>
 
 enum DSType : u8;
-enum Game : u16;
+enum Game : u32;
 enum Language : u8;
 class Profile5;
